+++ Assets/Scripts/Framework/BeatMapping/AudioWaveformDrawer.cs
