+++ Assets/Scripts/Framework/BeatMapping/LineData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac176253d460cd7449a25cdd9d37606b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
