+++ Assets/Scripts/Framework/BeatMapping/NewBehaviourScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe0d8de631aa25e42952d7990adcbb82
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
