+++ Assets/Scripts/Framework/BeatMapping/TImeStampData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af03446c2812f544c82763426aa4598f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
