+++ Assets/Scripts/Framework/BeatMapping/TimeStamper.cs
@@ -58,51 +58,6 @@ namespace ProefExamen.Framework.BeatMapping
 
         private AudioWaveformDrawer _waveformDrawer = null;
 
-        /// <summary>
-        /// Struct responsible for holding the necessary data for a gizmo line.
-        /// </summary>
-        [System.Serializable]
-        public struct LineData
-        {
-            public Vector2 startLinePoint;
-            public Vector2 endLinePoint;
-        }
-
-        /// <summary>
-        /// Class responsible for holding the necessary data for a time stamp.
-        /// </summary>
-        [System.Serializable]
-        public class TimeStampData
-        {
-            /// <summary>
-            /// Holds the start and end point of the time stamp.
-            /// </summary>
-            public LineData lineData;
-
-            /// <summary>
-            /// Holds the actual song time of the time stamp.
-            /// </summary>
-            public float songTime;
-
-            /// <summary>
-            /// Returns if this time stamp is selected.
-            /// </summary>
-            public bool isSelected;
-
-            /// <summary>
-            /// Constructor for the time stamp data.
-            /// </summary>
-            /// <param name="start">Start point of verticle line.</param>
-            /// <param name="end">End point of vertical line.</param>
-            /// <param name="time">Associated song time.</param>
-            public TimeStampData(Vector2 start, Vector2 end, float time)
-            {
-                lineData.startLinePoint = start;
-                lineData.endLinePoint = end;
-                songTime = time;
-            }
-        }
-
         private void Awake()
         {
             _waveformDrawer = FindObjectOfType<AudioWaveformDrawer>();
